Added staff curriculum hours fields
Added fields for assessing hours and coordination hours to the
input_pay_staff table. Will be used in calculting the cost of the
curriculum, by subtracting the hours worked in these ways (and
teaching) by Program Leaders/Curriculum Managers from the total
number of hours requried to fulfil the curriculum delivery.