Moved win32 import statement to function
win32com package causes errors on RtD and so is not
installed on RtD (i.e. not
in the documentation_requirements).
This causes an error in the
automodule functions for the
CLI docs when the package is imported
(in doc) command, hence moving the import statement to be
in-line, meaning it is only called when function is run.